Move actions panel JS code to webpack
