+++ app/javascript/packs/actionsPanel.js
@@ -1,0 +1,3 @@
+import { initializeModActionsMenu } from '../actionsPanel/actionsPanel';
+
+initializeModActionsMenu();
